svg/painting/parsing/marker-start-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing marker-start with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#MarkerStartProperty" rel="help"/><h:meta content="marker-start supports only the paint grammar 'none | marker-ref'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("marker-start", "auto");
test_invalid_value("marker-start", 'none url("https://example.com/")');

  ]]></script></svg>
